requirements.txt
requests==2.14.2
pycrypto==2.6.1
python-twitter==3.3
